refactor: add supertype conversion for narrow and widen conversion
